done HPF and RR

+++ scheduler.c
@@ -7,6 +7,7 @@
 #include <stdio.h>
 #include <signal.h>
 #include <unistd.h>
+#include <math.h>
 
 struct processData
 {
@@ -36,6 +37,7 @@ PCB createPCB(struct processData p)
 {
     PCB pcb;
     pcb.id = p.id;
+    pcb.pid = 0;
     pcb.arrival = p.arrivaltime;
     pcb.runtime = p.runningtime;
     pcb.remaining = p.runningtime;
@@ -55,18 +57,25 @@ PCB *getPCB(int id)
     return NULL;
 }
 
-void context_switch(PCB *old, PCB *new)
+void context_switch(FILE *pFile, PCB *old, PCB *new)
 {
-    if (old != NULL)
+    if (old != NULL && old->remaining > 0)
     {
-        printf("[TIME %d] STOP P%d\n", getClk(), old->id);
+        int waiting_time = getClk() - old->arrival - (old->runtime - old->remaining);
+        fprintf(pFile, "At\ttime\t%d\tprocess\t%d\tstopped\tarr\t%d\ttotal\t%d\tremain\t%d\twait\t%d\n", getClk(), old->id, old->arrival, old->runtime, old->remaining, waiting_time);
         kill(old->pid, SIGSTOP);
+        printf("[TIME %d] CONTEXT SWITCH (1 sec)\n", getClk());
+        sleep(1);
+    }
+    else if(old != NULL)
+    {
         printf("[TIME %d] CONTEXT SWITCH (1 sec)\n", getClk());
         sleep(1);
     }
     if (new->pid == 0)
     {
-        printf("[TIME %d] START P%d\n", getClk(), new->id);
+        int waiting_time = getClk() - new->arrival - (new->runtime - new->remaining);
+        fprintf(pFile, "At\ttime\t%d\tprocess\t%d\tstarted\tarr\t%d\ttotal\t%d\tremain\t%d\twait\t%d\n", getClk(), new->id, new->arrival, new->runtime, new->remaining, waiting_time);
         int pid = fork();
         if (pid == 0)
         {
@@ -81,12 +90,13 @@ void context_switch(PCB *old, PCB *new)
     }
     else
     {
-        printf("[TIME %d] RESUME P%d\n", getClk(), new->id);
+        int waiting_time = getClk() - new->arrival - (new->runtime - new->remaining);
+        fprintf(pFile, "At\ttime\t%d\tprocess\t%d\tresumed\tarr\t%d\ttotal\t%d\tremain\t%d\twait\t%d\n", getClk(), new->id, new->arrival, new->runtime, new->remaining, waiting_time);
         kill(new->pid, SIGCONT);
     }
 }
 
-void HPF()
+void HPF(FILE *pFile)
 {
     struct msgbuff message;
     ready_queue = createQueue();
@@ -105,6 +115,12 @@ void HPF()
             {
                 printf("[TIME %d] P%d finished immediately\n", getClk(), pcb.id);
                 finished_processes++;
+                int waiting_time = getClk() - current->arrival - current->runtime;
+                int TA = getClk() - current->arrival;
+                float WTA = round(((float)TA / current->runtime) * 100) / 100;
+                current->waiting_time = waiting_time;
+                current->WTA = WTA;
+                fprintf(pFile, "At\ttime\t%d\tprocess\t%d\tfinished\tarr\t%d\ttotal\t%d\tremain\t%d\twait\t%d\tTA\t%d\tWTA\t%.2f\n", getClk(), current->id, current->arrival, current->runtime, current->remaining, waiting_time, TA, WTA);
                 continue;
             }
             pcbs[pcb_count++] = pcb;
@@ -120,7 +136,7 @@ void HPF()
                 enqueuePri(ready_queue, current->id, current->priority);
                 int id = dequeue(ready_queue);
                 PCB *next = getPCB(id);
-                context_switch(current, next);
+                context_switch(pFile, current, next);
                 current = next;
             }
         }
@@ -129,7 +145,7 @@ void HPF()
             int id = dequeue(ready_queue);
             PCB *next = getPCB(id);
             printf("[TIME %d] Pick P%d from queue\n", getClk(), id);
-            context_switch(NULL, next);
+            context_switch(pFile, NULL, next);
             current = next;
         }
         if (current != NULL)
@@ -142,12 +158,24 @@ void HPF()
             current->remaining--;
             if (current->remaining == 0)
             {
-                printf("[TIME %d] FINISH P%d\n",
-                       getClk(),
-                       current->id);
+                PCB *old = current;
                 waitpid(current->pid, NULL, 0);
                 finished_processes++;
+                int waiting_time = getClk() - current->arrival - current->runtime;
+                int TA = getClk() - current->arrival;
+                float WTA = round(((float)TA / current->runtime) * 100) / 100;
+                current->waiting_time = waiting_time;
+                current->WTA = WTA;
+                fprintf(pFile, "At\ttime\t%d\tprocess\t%d\tfinished\tarr\t%d\ttotal\t%d\tremain\t%d\twait\t%d\tTA\t%d\tWTA\t%.2f\n", getClk(), current->id, current->arrival, current->runtime, current->remaining, waiting_time, TA, WTA);
+
                 current = NULL;
+                if (!isEmpty(ready_queue))
+                {
+                    int id = dequeue(ready_queue);
+                    PCB *next = getPCB(id);
+                    context_switch(pFile, old, next);
+                    current = next;
+                }
             }
         }
     }
@@ -158,7 +186,7 @@ void HPF()
 // ROUND ROBIN ALGORITHM
 ////////////////////////////////////////////////////////////
 
-void RR(int quantum)
+void RR(FILE *pFile, int quantum)
 {
     struct msgbuff message;
     ready_queue = createQueue();
@@ -181,6 +209,12 @@ void RR(int quantum)
             {
                 printf("[TIME %d] P%d finished immediately\n", getClk(), pcb.id);
                 finished_processes++;
+                int waiting_time = getClk() - current->arrival - current->runtime;
+                int TA = getClk() - current->arrival;
+                float WTA = round(((float)TA / current->runtime) * 100) / 100;
+                current->waiting_time = waiting_time;
+                current->WTA = WTA;
+                fprintf(pFile, "At\ttime\t%d\tprocess\t%d\tfinished\tarr\t%d\ttotal\t%d\tremain\t%d\twait\t%d\tTA\t%d\tWTA\t%.2f\n", getClk(), current->id, current->arrival, current->runtime, current->remaining, waiting_time, TA, WTA);
                 continue;
             }
 
@@ -196,7 +230,7 @@ void RR(int quantum)
 
             printf("[TIME %d] Pick P%d\n", getClk(), id);
 
-            context_switch(NULL, next);
+            context_switch(pFile, NULL, next);
             current = next;
             quantum_counter = 0;
         }
@@ -218,14 +252,28 @@ void RR(int quantum)
             // ===== FINISH =====
             if (current->remaining == 0)
             {
+                PCB *old = current;
                 printf("[TIME %d] FINISH P%d\n",
                        getClk(), current->id);
 
                 waitpid(current->pid, NULL, 0);
 
                 finished_processes++;
+                int waiting_time = getClk() - current->arrival - current->runtime;
+                int TA = getClk() - current->arrival;
+                float WTA = round(((float)TA / current->runtime) * 100) / 100;
+                current->waiting_time = waiting_time;
+                current->WTA = WTA;
+                fprintf(pFile, "At\ttime\t%d\tprocess\t%d\tfinished\tarr\t%d\ttotal\t%d\tremain\t%d\twait\t%d\tTA\t%d\tWTA\t%.2f\n", getClk(), current->id, current->arrival, current->runtime, current->remaining, waiting_time, TA, WTA);
                 current = NULL;
                 quantum_counter = 0;
+                if (!isEmpty(ready_queue))
+                {
+                    int id = dequeue(ready_queue);
+                    PCB *next = getPCB(id);
+                    context_switch(pFile, old, next);
+                    current = next;
+                }
             }
 
             // ===== QUANTUM EXPIRE =====
@@ -245,7 +293,7 @@ void RR(int quantum)
 
                     quantum_counter = 0;
 
-                    context_switch(old, next);
+                    context_switch(pFile, old, next);
                     current = next;
                 }
                 else
@@ -282,23 +330,51 @@ int main(int argc, char *argv[])
 
     msg_id = msgget(MSGKEY, IPC_CREAT | 0666);
 
+    FILE *pFile;
+    pFile = fopen("schedulerLog.txt", "w");
+    fprintf(pFile, "#At\ttime\tx\tprocess\ty\tstate\tarr\tw\ttotal\tz\tremain\ty\twait\tk\n");
+
     // FIX: removed the SIGCHLD handler — it was racing with our accounting logic.
     // Child cleanup is now done explicitly with waitpid() when remaining hits 0.
     if (algo == 1)
     {
         // HPF
-        HPF();
+        HPF(pFile);
     }
     else if (algo == 2)
     {
         // RR
-        RR(quantum);
+        RR(pFile, quantum);
     }
     else
     {
         printf("Algorithm %d not implemented in this snippet.\n", algo);
     }
-
+    fclose(pFile);
+    int total_waiting_time = 0;
+    float total_WTA_time = 0;
+    int total_runtime = 0;
+    float std_WTA = 0;
+    for (int i = 0; i < number_of_processes; i++)
+    {
+        total_waiting_time += pcbs[i].waiting_time;
+        total_WTA_time += pcbs[i].WTA;
+        total_runtime += pcbs[i].runtime;
+    }
+    float avg_waiting_time = round(((float)total_waiting_time / number_of_processes) * 100) / 100;
+    float avg_WTA_time = round(((float)total_WTA_time / number_of_processes) * 100) / 100;
+    float cpu_utilization = round(((float)total_runtime / getClk()) * 10000) / 100;
+    for (int i = 0; i < number_of_processes; i++)
+    {
+        std_WTA += pow(pcbs[i].WTA - avg_WTA_time, 2);
+    }
+    std_WTA = round((sqrt(std_WTA / number_of_processes)) * 100) / 100;
+    pFile = fopen("schedulerPerf.txt", "w");
+    fprintf(pFile, "CPU utilization = %.2f%%\n", cpu_utilization);
+    fprintf(pFile, "Avg WTA = %.2f\n", avg_WTA_time);
+    fprintf(pFile, "Avg Waiting = %.2f\n", avg_waiting_time);
+    fprintf(pFile, "Std WTA = %.2f\n", std_WTA);
+    fclose(pFile);
     printf("All processes finished. Cleaning up...\n");
     destroyClk(true);
     return 0;
